Lib

Documentation

chooseWords :: [Text] -> [Int] -> Int -> IO [String]

Random word collector. Creates a list of random words with the help of a list of random numbers which should be cryptographicaly secure if used for a password. Usage: chooseWords [random word list] [list of Ints] (length of the word list)

rand :: Int -> IO [Int]

Auxiliary function for random numbers. Creates a list of random numbers which size is the input from a cryptographically secure random number generator (raaz).

password :: String -> String -> [String] -> String

Password creation function. Two Strings get interlaced between a list of Strings to form a secure easy to remember password.